cardinal-grip-s3/requirements.txt
ajsonrpc==1.2.0
anyio==4.12.0
argcomplete==3.6.3
bleak==1.1.1
bottle==0.13.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
h11==0.16.0
idna==3.11
marshmallow==3.26.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pipx==1.8.0
platformdirs==4.5.0
platformio==6.1.18
pyelftools==0.32
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreBluetooth==12.1
pyobjc-framework-libdispatch==12.1
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
semantic-version==2.10.0
six==1.17.0
starlette==0.46.2
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.34.3
websocket-client==1.9.0
wsproto==1.3.2